+++ packages/expo-cellular/android/src/main/java/expo/modules/cellular/CellularGeneration.kt
@@ -1,8 +1,8 @@
 package expo.modules.cellular
 
-import expo.modules.kotlin.types.EnumArgument
+import expo.modules.kotlin.types.Enumerable
 
-enum class CellularGeneration(val value: Int) : EnumArgument {
+enum class CellularGeneration(val value: Int) : Enumerable {
   UNKNOWN(0),
   CG_2G(1),
   CG_3G(2),