+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/Enumerable.kt
@@ -6,4 +6,4 @@ import expo.modules.core.interfaces.DoNotStrip
  * A interface that allows converting raw values to enum cases.
  */
 @DoNotStrip
-interface EnumArgument
+interface Enumerable